pyproject.toml
[build-system]
requires = ["poetry>=1.5.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "my_package"
version = "0.1.0"
description = "Whisper Open-ai Media GUI"
authors = ["Kevin James <krjii@indyhustles.com>"]
license = "BSD-3-Clause"
readme = "README.rst"
keywords = ["Whipser", "Open-AI"]
classifiers = [
    "Programming Language :: Python :: 3.9",
    "License :: OSI Approved :: BSD License",
    "Intended Audience :: Developers",
]

[tool.poetry.dependencies]
python = ">=3.9,<3.14"
openai-whisper = "*"
setuptools-rust = "*"
pyside6 = "*"

[tool.poetry.scripts]
my-script = "my_package.module:function"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
